regimeflex/requirements.txt
colorama>=0.4.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0



